refactor dynamicops folder
