+++ mmrazor/models/architectures/dynamic_ops/__init__.py
@@ -1,22 +1,15 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .dynamic_conv import BigNasConv2d, DynamicConv2d, OFAConv2d
-from .dynamic_linear import DynamicLinear
-from .dynamic_mixins import (DynamicBatchNormMixin, DynamicChannelMixin,
-                             DynamicLinearMixin, DynamicMixin)
-from .dynamic_norm import (DynamicBatchNorm1d, DynamicBatchNorm2d,
-                           DynamicBatchNorm3d, SwitchableBatchNorm2d)
+from .bricks.dynamic_conv import BigNasConv2d, DynamicConv2d, OFAConv2d
+from .bricks.dynamic_linear import DynamicLinear
+from .bricks.dynamic_norm import (DynamicBatchNorm1d, DynamicBatchNorm2d,
+                                  DynamicBatchNorm3d, SwitchableBatchNorm2d)
+from .mixins.dynamic_conv_mixins import DynamicConvMixin
+from .mixins.dynamic_mixins import (DynamicBatchNormMixin, DynamicChannelMixin,
+                                    DynamicLinearMixin, DynamicMixin)
 
 __all__ = [
-    'BigNasConv2d',
-    'DynamicConv2d',
-    'OFAConv2d',
-    'DynamicLinear',
-    'DynamicBatchNorm1d',
-    'DynamicBatchNorm2d',
-    'DynamicBatchNorm3d',
-    'DynamicMixin',
-    'DynamicChannelMixin',
-    'DynamicBatchNormMixin',
-    'DynamicLinearMixin',
-    'SwitchableBatchNorm2d',
+    'BigNasConv2d', 'DynamicConv2d', 'OFAConv2d', 'DynamicLinear',
+    'DynamicBatchNorm1d', 'DynamicBatchNorm2d', 'DynamicBatchNorm3d',
+    'DynamicMixin', 'DynamicChannelMixin', 'DynamicBatchNormMixin',
+    'DynamicLinearMixin', 'SwitchableBatchNorm2d', 'DynamicConvMixin'
 ]
