+++ mmrazor/models/mutables/mutable_channel/groups/channel_group.py
@@ -17,8 +17,7 @@ from typing import Any, Dict, List, Tuple, TypeVar, Union
 import torch.nn as nn
 from torch.nn import Module
 
-from mmrazor.models.architectures.dynamic_ops.dynamic_mixins import \
-    DynamicChannelMixin
+from mmrazor.models.architectures.dynamic_ops.mixins import DynamicChannelMixin
 from mmrazor.structures.graph import ModuleGraph, ModuleNode
 from mmrazor.utils import IndexDict
 from ..base_mutable_channel import BaseMutableChannel
