+++ tests/test_models/test_architectures/test_dynamic_op/test_bricks/test_dynamic_conv.py
@@ -226,7 +226,8 @@ def test_kernel_dynamic_conv2d(dynamic_class: Type[nn.Conv2d],
 @pytest.mark.parametrize('dynamic_class', [OFAConv2d, BigNasConv2d])
 def test_mutable_kernel_dynamic_conv2d_grad(
         dynamic_class: Type[nn.Conv2d]) -> None:
-    from mmrazor.models.architectures.dynamic_ops import dynamic_conv_mixins
+    from mmrazor.models.architectures.dynamic_ops.mixins import \
+        dynamic_conv_mixins
 
     kernel_size_list = [3, 5, 7]
     d_conv2d = dynamic_class(
